+++ src/main/java/com/example/usan/controller/StorageController.java
@@ -45,7 +45,7 @@ public class StorageController {
     }
 
     @GetMapping("/{id}")
-    public String findById(@PathVariable int id, Model model) {
+    public String findById(@PathVariable Long id, Model model) {
         model.addAttribute("storage", storageService.sto_detail(id));
         model.addAttribute("umbrella", umbrellaService.umb_upload());
         return "storage/storage_detail";
@@ -55,7 +55,7 @@ public class StorageController {
 
 
     @GetMapping("/admin/{id}")
-    public String findById_ADMIN(@PathVariable int id, Model model) {
+    public String findById_ADMIN(@PathVariable Long id, Model model) {
         model.addAttribute("storage", storageService.sto_detail(id));
         model.addAttribute("umbrella", umbrellaService.umb_upload());
         return "storage/storage_detail_ADMIN";
