+++ src/main/java/com/example/usan/dto/ReplySaveRequestDto.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReplySaveRequestDto {
-    private int userId;
-    private int boardId;
+    private Long userId;
+    private Long boardId;
     private String content;
 }
